refactor(editor): remove legacy imperative dialog helpers, extract builder utils
- Delete unused backward-compat exports from expression-dialog.ts:
  renderFieldPaths, createFieldPathItem, renderQuickReference, updatePreview.
  No callers existed outside the file.
- Extract pure expression builder helpers into new
  expression-builder.ts: parseFormatDateExpression, wrapFormatDate,
  BuilderState, tryParseAsBuilderExpression, isStaleFieldReference,
  buildExpression.
- expression-dialog.ts now only exports the dialog options/result types,
  the openExpressionDialog wrapper, and JSONATA_QUICK_REFERENCE.
- Fix duplicate HTMLElementTagNameMap declaration in EpExpressionDialog.ts.
- Update imports in EpExpressionDialog.ts and expression-dialog.test.ts.
